AI Agent/requirements.txt
langchain==0.1.20
langchain-community
langchain-openai
langchain-anthropic
wikipedia
python-dotenv
pydantic<2
openai
anthropic
